urlshortener/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
drf-yasg==1.21.7
django-ratelimit==4.1.0
django-password-validators==1.3.0
django-simple-history==3.4.0 